+++ .editorconfig
@@ -1,5 +1,3 @@
-root = true
-
 [*]
 indent_style = space
 indent_size = 4
