+++ .jshintrc
@@ -13,10 +13,5 @@
     "unused": true,
     "browser": true,
     "latedef": false,
-    "globals": {
-        "_": true,
-        "define": true,
-        "root": true,
-        "factory": true
-    }
+    "esversion": 6
 }
