+++ src/modifiers/index.js
@@ -1,0 +1,19 @@
+import applyStyle from './applyStyle';
+import arrow from './arrow';
+import flip from './flip';
+import keepTogether from './keepTogether';
+import offset from './offset';
+import preventOverflow from './preventOverflow';
+import shift from './shift';
+
+const modifiers = {
+    applyStyle,
+    arrow,
+    flip,
+    keepTogether,
+    offset,
+    preventOverflow,
+    shift
+};
+
+export default modifiers;
