+++ src/modifiers/keepTogether.js
@@ -1,0 +1,29 @@
+import getPopperClientRect from '../utils/getPopperClientRect';
+
+/**
+ * Modifier used to make sure the popper is always near its reference
+ * @method
+ * @memberof Popper.modifiers
+ * @argument {Object} data - The data object generated by _update method
+ * @returns {Object} The data object, properly modified
+ */
+export default function keepTogether(data) {
+    var popper  = getPopperClientRect(data.offsets.popper);
+    var reference = data.offsets.reference;
+    var f = Math.floor;
+
+    if (popper.right < f(reference.left)) {
+        data.offsets.popper.left = f(reference.left) - popper.width;
+    }
+    if (popper.left > f(reference.right)) {
+        data.offsets.popper.left = f(reference.right);
+    }
+    if (popper.bottom < f(reference.top)) {
+        data.offsets.popper.top = f(reference.top) - popper.height;
+    }
+    if (popper.top > f(reference.bottom)) {
+        data.offsets.popper.top = f(reference.bottom);
+    }
+
+    return data;
+}
