+++ src/modifiers/offset.js
@@ -1,0 +1,26 @@
+/**
+ * Modifier used to add an offset to the popper, useful if you more granularity positioning your popper.
+ * The offsets will shift the popper on the side of its reference element.
+ * @method
+ * @memberof Popper.modifiers
+ * @argument {Object} data - The data object generated by _update method
+ * @returns {Object} The data object, properly modified
+ */
+export default function offset(data) {
+    var offset = data.instance._options.offset;
+    var popper  = data.offsets.popper;
+
+    if (data.placement.indexOf('left') !== -1) {
+        popper.top -= offset;
+    }
+    else if (data.placement.indexOf('right') !== -1) {
+        popper.top += offset;
+    }
+    else if (data.placement.indexOf('top') !== -1) {
+        popper.left -= offset;
+    }
+    else if (data.placement.indexOf('bottom') !== -1) {
+        popper.left += offset;
+    }
+    return data;
+}
