+++ src/modifiers/preventOverflow.js
@@ -1,0 +1,50 @@
+import getPopperClientRect from '../utils/getPopperClientRect';
+
+/**
+ * Modifier used to make sure the popper does not overflows from it's boundaries
+ * @method
+ * @memberof Popper.modifiers
+ * @argument {Object} data - The data object generated by `update` method
+ * @returns {Object} The data object, properly modified
+ */
+export default function preventOverflow(data) {
+    var order = data.instance._options.preventOverflowOrder;
+    var popper = getPopperClientRect(data.offsets.popper);
+
+    var check = {
+        left: function() {
+            var left = popper.left;
+            if (popper.left < data.boundaries.left) {
+                left = Math.max(popper.left, data.boundaries.left);
+            }
+            return { left: left };
+        },
+        right: function() {
+            var left = popper.left;
+            if (popper.right > data.boundaries.right) {
+                left = Math.min(popper.left, data.boundaries.right - popper.width);
+            }
+            return { left: left };
+        },
+        top: function() {
+            var top = popper.top;
+            if (popper.top < data.boundaries.top) {
+                top = Math.max(popper.top, data.boundaries.top);
+            }
+            return { top: top };
+        },
+        bottom: function() {
+            var top = popper.top;
+            if (popper.bottom > data.boundaries.bottom) {
+                top = Math.min(popper.top, data.boundaries.bottom - popper.height);
+            }
+            return { top: top };
+        }
+    };
+
+    order.forEach(function(direction) {
+        data.offsets.popper = Object.assign(popper, check[direction]());
+    });
+
+    return data;
+}
