+++ src/modifiers/shift.js
@@ -1,0 +1,37 @@
+import getPopperClientRect from '../utils/getPopperClientRect';
+
+/**
+ * Modifier used to shift the popper on the start or end of its reference element side
+ * @method
+ * @memberof Popper.modifiers
+ * @argument {Object} data - The data object generated by `update` method
+ * @returns {Object} The data object, properly modified
+ */
+export default function shift(data) {
+    var placement = data.placement;
+    var basePlacement = placement.split('-')[0];
+    var shiftVariation = placement.split('-')[1];
+
+    // if shift shiftVariation is specified, run the modifier
+    if (shiftVariation) {
+        var reference = data.offsets.reference;
+        var popper = getPopperClientRect(data.offsets.popper);
+
+        var shiftOffsets = {
+            y: {
+                start:  { top: reference.top },
+                end:    { top: reference.top + reference.height - popper.height }
+            },
+            x: {
+                start:  { left: reference.left },
+                end:    { left: reference.left + reference.width - popper.width }
+            }
+        };
+
+        var axis = ['bottom', 'top'].indexOf(basePlacement) !== -1 ? 'x' : 'y';
+
+        data.offsets.popper = Object.assign(popper, shiftOffsets[axis][shiftVariation]);
+    }
+
+    return data;
+}
