+++ src/utils/getBoundingClientRect.js
@@ -1,0 +1,18 @@
+/**
+ * Get bounding client rect of given element
+ * @function
+ * @ignore
+ * @param {HTMLElement} element
+ * @return {Object} client rect
+ */
+export default function getBoundingClientRect(element) {
+    var rect = element.getBoundingClientRect();
+    return {
+        left: rect.left,
+        top: rect.top,
+        right: rect.right,
+        bottom: rect.bottom,
+        width: rect.right - rect.left,
+        height: rect.bottom - rect.top
+    };
+}
