+++ src/utils/getPopperClientRect.js
@@ -1,0 +1,13 @@
+/**
+ * Given the popper offsets, generate an output similar to getBoundingClientRect
+ * @function
+ * @ignore
+ * @argument {Object} popperOffsets
+ * @returns {Object} ClientRect like output
+ */
+export default function getPopperClientRect(popperOffsets) {
+    var offsets = Object.assign({}, popperOffsets);
+    offsets.right = offsets.left + offsets.width;
+    offsets.bottom = offsets.top + offsets.height;
+    return offsets;
+}
