+++ src/utils/isFixed.js
@@ -1,0 +1,19 @@
+import getStyleComputedProperty from './getStyleComputedProperty';
+
+/**
+ * Check if the given element is fixed or is inside a fixed parent
+ * @function
+ * @ignore
+ * @argument {Element} element
+ * @argument {Element} customContainer
+ * @returns {Boolean} answer to "isFixed?"
+ */
+export default function isFixed(element) {
+    if (element === window.document.body) {
+        return false;
+    }
+    if (getStyleComputedProperty(element, 'position') === 'fixed') {
+        return true;
+    }
+    return element.parentNode ? isFixed(element.parentNode) : element;
+}
