+++ src/utils/isModifierRequired.js
@@ -1,0 +1,16 @@
+/**
+ * Helper used to know if the given modifier depends from another one.
+ * @method
+ * @memberof Popper
+ * @returns {Boolean}
+ */
+export default function isModifierRequired(options, requesting, requested) {
+    return options.modifiers.find(modifier => {
+        if (modifier.name === requested) {
+            return true;
+        } else if (modifier.name === requesting) {
+            return false;
+        }
+        return false;
+    }) !== undefined;
+}
